commands: fix required argument for resource
keep the same behavior like other commands; when a resource must be
supplied, display usage info if not provided.
I made an exception on docker since if we do not provide any args, the
docker usage is displayed.

Arguments are required by default in Click, but when no default is
provided, it tests the nargs value for > 0, as we use nargs=-1 for
unlimited number of resources, this test failed.

Signed-off-by: William Dauchy <william@gandi.net>