iaas: do not output API result when creating a VM in the background
`gandi vm create --background ...` pretty-prints the result of the
`hosting.vm.create_from` API call, which is a list of operations.
This is confusing to the user (it looks like the CLI is just spitting
a bunch of debugging output). I propose to display an informative
message instead. This makes the CLI nicer to use in automation
scripts.