Accept the fact that vm create can be only with a vlan.
If you give --vlan but no --ip-version, it will create a vm without public iface.
If you give --vlan and --ip-version, it will create a vm with a public iface and then attach a private one.
If you give --ip-version and no --vlan, it will create a vm with a public iface only.