iaas, paas: change default prefix for auto generated names
we are used to let the CLI decide for a random name in order to create
the resource.
sometimes the later is not temporary, we just wanted a generated name.
this make more sense to name it starting from the type of resource.

moreover when listing resources such as VMs and disks, it's sometimes
confusing to write an attach command with similar names, e.g:

$ gandi disk attach tempXXXXA tempXXXXZ

(did I put the right resource names?)

instead of:

$ gandi disk attach vdiXXXXA vmXXXXZ

which is a bit clearer to me.

Signed-off-by: William Dauchy <william@gandi.net>