core, utils: adding ascii_sparks
the code is imported from https://pypi.python.org/pypi/ascii_sparks/
we first thought to add it through the setup.py file since it's quite
easy to add new import when using pypi; but if you are using the CLI
with a package from your distribution, and the dependency is not
available, you will not be able to use the functionality. Since this
particular lib does not seem to be packaged at the present time by any
well known distributions, we choose to copy it in our repository.

this will allow to display simple stats from metrics in info commands
the plan is also to be able to query metrics and pipe results in an
external program

Acked-by: Dejan Filipovic <sayoun@gandi.net>
Signed-off-by: William Dauchy <william@gandi.net>