vm: add stat option in `info` command
this will display bare and simple statistics of the network vm
the idea is to have a general view of the vm workload

this has been added as an option since it makes the info command slower
than usual.
if no stat are found (zero), the graph will be blank

in the future, we also can add specifics stats in `ip info` for each
interface.

Signed-off-by: William Dauchy <william@gandi.net>