vm: delete: Fix delete when we reach the list limit
If we try to delete a vm that isn't present in the default limit of
gandi.iaas.list, the script fails.
Instead of falling back to none when iterating on the list, fallback to
gandi.iaas.info
The resource have already been checked so we know that we will have a right
object.

Signed-off-by: Emmanuel Vadot <emmanuel.vadot@gandi.net>